Rename getResourceStrings -> getTemplateStrings
