+++ plugin-hrm-form/src/___tests__/mockGetConfig.ts
@@ -56,7 +56,7 @@ const mockGetResourceStrings = jest.fn(() => baseMockConfig.strings);
 jest.mock('../hrmConfig', () => ({
   getHrmConfig: mockGetConfig,
   getAseloFeatureFlags: mockGetAseloFeatureFlags,
-  getResourceStrings: mockGetResourceStrings,
+  getTemplateStrings: mockGetResourceStrings,
   initializeConfig: jest.fn(),
   getDefinitionVersions: jest.fn(),
 }));
