+++ plugin-hrm-form/src/components/case/casePrint/CasePrintSummary.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 import { Text, View } from '@react-pdf/renderer';
 
 import styles from './styles';
-import { getResourceStrings } from '../../../hrmConfig';
+import { getTemplateStrings } from '../../../hrmConfig';
 
 type OwnProps = {
   summary?: string;
@@ -30,7 +30,7 @@ type OwnProps = {
 type Props = OwnProps;
 
 const CasePrintSummary: React.FC<Props> = ({ summary }) => {
-  const strings = getResourceStrings();
+  const strings = getTemplateStrings();
 
   return (
     <View>
