+++ plugin-hrm-form/src/components/caseList/filters/Filters.tsx
@@ -41,7 +41,7 @@ import {
 import CategoriesFilter, { Category } from './CategoriesFilter';
 import { caseListBase, configurationBase, namespace, RootState } from '../../../states';
 import * as CaseListSettingsActions from '../../../states/caseList/settings';
-import { getAseloFeatureFlags, getHrmConfig, getResourceStrings } from '../../../hrmConfig';
+import { getAseloFeatureFlags, getHrmConfig, getTemplateStrings } from '../../../hrmConfig';
 /**
  * Reads the definition version and returns and array of items (type Item[])
  * to be used as the options for the status filter
@@ -167,7 +167,7 @@ const Filters: React.FC<Props> = ({
   updateCaseListFilter,
   clearCaseListFilter,
 }) => {
-  const strings = getResourceStrings();
+  const strings = getTemplateStrings();
   const featureFlags = getAseloFeatureFlags();
   const { helpline } = getHrmConfig();
